[SPARK-28084][SPARK-SQL]Resolving the partition column name based on the resolver which follows spark.sql.caseSensitive property configuration value. What changes were proposed in this pull request? LOAD DATA command resolves the partition column name as case sensitive manner, where as in insert command the partition column name will be resolved using the SQLConf resolver where the names will be  resolved  based on spark.sql.caseSensitive property. same logic can be applied for resolving the partition column names in LOAD COMMAND.
How was this patch tested?
Existing UT and manual testing.
